[add, fix] マイグレーション、シーダー、モデルの変更

+++ database/migrations/2020_04_21_052746_create_products_table.php
@@ -14,8 +14,8 @@ class CreateProductsTable extends Migration
     public function up()
     {
         Schema::create('products', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->bigInteger('subject_id')->unsigned()->comment('科目ID');
+            $table->increments('id');
+            $table->integer('subject_id')->unsigned()->comment('科目ID');
             $table->string('product_name')->comment('商品名');
             $table->integer('price')->comment('金額');
             $table->integer('sort_no')->comment('ソートNo');
